moved shared mocks to shared folder
