+++ bridge/server/test/service-states.spec.ts
@@ -1,6 +1,6 @@
 import request from 'supertest';
 import MockAdapter from 'axios-mock-adapter';
-import { ProjectQualityGatesResponse, ProjectResponse } from '../fixtures/project-response.mock';
+import { ProjectQualityGatesResponse, ProjectResponse } from '../../shared/fixtures/project-response.mock';
 import { OpenRemediationsResponse } from '../../shared/fixtures/open-remediations-response.mock';
 import {
   ServiceStateQualityGatesOnlyResponse,
