+++ bridge/shared/fixtures/project-response.mock.ts
