+++ bridge/shared/fixtures/service-deployment-response.mock.ts
@@ -1,4 +1,4 @@
-import { EvaluationFinishedStagingResponse } from '../../server/fixtures/traces-response.mock';
+import { EvaluationFinishedStagingResponse } from './traces-response.mock';
 
 const serviceDeploymentMock = {
   state: 'finished',
